src/producer/requirements.txt
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
boto3==1.34.0
schedule==1.2.0 